+++ src/FloatSoda/FloatSodaApp.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Concurrent;
 using FloatSoda.Geometrics;
 using FloatSoda.Render;
+using FloatSoda.Render.Layout;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using OVRSharp;
